+++ R/post_normal_covar_const.R
@@ -7,6 +7,7 @@
 #' @param u_omega_i matrix of error variances of the measurement equation.
 #' Either a \eqn{K \times K} matrix for constant variances or
 #' a \eqn{KT \times KT} matrix for time varying variances.
+#' @param k number of endogenous variables.
 #' @param prior_mean vector of prior means. In case of TVP, this vector is used
 #' as initial condition.
 #' @param prior_covariance_i inverse prior covariance matrix. In case of TVP, this matrix
@@ -28,46 +29,33 @@
 #' # Load example data
 #' data("e1")
 #' y <- log(t(e1))
+#' k <- nrow(y)
+#' y <- matrix(y)
 #' 
 #' # Generate artificial draws of other matrices
-#' u_omega_i <- diag(1, 3)
+#' u_omega_i <- Matrix(diag(1, k))
 #' prior_mean <- matrix(0, 3)
-#' prior_covariance_i <- diag(0, 3)
+#' prior_covariance_i <- Matrix(diag(0, 3))
 #' 
 #' # Obtain posterior draw
-#' post_normal_covar_const(y, u_omega_i, prior_mean, prior_covariance_i)
+#' post_normal_covar_const(y, u_omega_i, k, prior_mean, prior_covariance_i)
 #'
 #' @export
-post_normal_covar_const <- function(y, u_omega_i, prior_mean, prior_covariance_i) {
+post_normal_covar_const <- function(y, u_omega_i, k, prior_mean, prior_covariance_i) {
   
-  k <- nrow(y)
   if (k == 1L) {
     stop("Argument 'y' must contain at least two variables.")
   }
   n_covar <- k * (k - 1) / 2
-  tt <- ncol(y)
-  y <- matrix(y)
+  tt <- nrow(y) / k
   
   # Generate z for lower triangular design
   # Use C++ to speed up the for loop
-  z <- .prep_covar_data(y, k, tt, FALSE)
-  
-  # Get positions of values, on which variables in z are regressed
-  pos_used <- rep(k * 0:(tt - 1), each = k - 1) + 2:k
-  
-  # Trim endogenous variables
-  y <- matrix(y[pos_used, ])
-  
-  # Adapt error variance matrix
-  if (NCOL(u_omega_i) == k & NCOL(u_omega_i) == k) {
-    u_omega_i <- kronecker(Matrix::Diagonal(tt, 1), u_omega_i)
-  }
-  # Trim error variance matrix
-  u_omega_i <- u_omega_i[pos_used, pos_used]
+  temp <- covar_prepare_data(y, u_omega_i, k, tt, FALSE)
   
   # Draw coefficients
-  v_post <- prior_covariance_i + crossprod(z, u_omega_i) %*% z
-  mu_post <- solve(v_post, prior_covariance_i %*% prior_mean + crossprod(z, u_omega_i) %*% y)
+  v_post <- prior_covariance_i + crossprod(temp$z, temp$omega_i) %*% temp$z
+  mu_post <- solve(v_post, prior_covariance_i %*% prior_mean + crossprod(temp$z, temp$omega_i) %*% temp$y)
   
   return(matrix(mu_post + solve(chol(v_post), matrix(stats::rnorm(n_covar)))))
 }